+++ packages/wave/src/util/deploy/index.ts
@@ -1,3 +1,4 @@
+import dns from "node:dns";
 import { errorToString } from "@vastly/utils";
 import { Client } from "../client.js";
 import { isAPIError } from "../error.js";
@@ -18,18 +19,21 @@ export default async function doDeploy(client: Client, stage: string) {
     return await executeDeploy(client, stage);
   } catch (err: unknown) {
     output.error(errorToString(err));
-    return 0;
+    return;
   }
 }
 
 async function executeDeploy(client: Client, stage: string): Promise<DeployResult> {
   const { apiUrl } = client;
+  dns.setDefaultResultOrder("ipv4first");
 
   try {
     // hardcoded account_id (stage) to infra-prod
     return await client.fetch<DeployResult>(`http://localhost:4000/prod/deploy/908170539157`, {
       method: "POST",
-      headers: {},
+      headers: {
+        // Authorization: client.config.token,
+      },
     });
   } catch (err: unknown) {
     if (isAPIError(err)) {