+++ services/deploy/serverless.ts
@@ -69,161 +69,14 @@ const serverlessConfiguration: AWS = {
           },
           Policies: [
             {
-              PolicyName: "DeploymentPolicy",
+              PolicyName: "SSTDeployPolicy",
               PolicyDocument: {
                 Version: "2012-10-17",
                 Statement: [
-                  //  Permissions to deploy the bootstrap stack once for each AWS account
                   {
-                    Sid: "AllowCDKManageToolkitStack",
                     Effect: "Allow",
-                    Action: [
-                      "cloudformation:CreateChangeSet",
-                      "cloudformation:DeleteChangeSet",
-                      "cloudformation:DeleteStack",
-                      "cloudformation:DescribeChangeSet",
-                      "cloudformation:DescribeStacks",
-                      "cloudformation:DescribeStackEvents",
-                      "cloudformation:ExecuteChangeSet",
-                      "cloudformation:GetTemplate",
-                    ],
-                    Resource: ["arn:aws:cloudformation:us-east-1:112233445566:stack/CDKToolkit/*"],
-                  },
-                  // Permissions to create the CDK bootstrap roles
-                  {
-                    Sid: "AllowCDKManageToolkitRoles",
-                    Effect: "Allow",
-                    Action: [
-                      "iam:AttachRolePolicy",
-                      "iam:CreateRole",
-                      "iam:DeleteRole",
-                      "iam:DeleteRolePolicy",
-                      "iam:DetachRolePolicy",
-                      "iam:GetRole",
-                      "iam:GetRolePolicy",
-                      "iam:PutRolePolicy",
-                      "iam:TagRole",
-                    ],
-                    Resource: [
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-cfn-exec-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-file-publishing-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-image-publishing-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-lookup-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-deploy-role-*",
-                    ],
-                  },
-                  // Permissions to create the CDK bootstrap bucket
-                  {
-                    Sid: "AllowCDKManageToolkitBucket",
-                    Effect: "Allow",
-                    Action: [
-                      "s3:CreateBucket",
-                      "s3:DeleteBucketPolicy",
-                      "s3:GetEncryptionConfiguration",
-                      "s3:GetBucketPolicy",
-                      "s3:PutBucketPolicy",
-                      "s3:PutBucketVersioning",
-                      "s3:PutEncryptionConfiguration",
-                      "s3:PutLifecycleConfiguration",
-                      "s3:PutBucketPublicAccessBlock",
-                    ],
-                    Resource: ["arn:aws:s3:::cdk-hnb659fds-assets-*"],
-                  },
-                  // Permissions to create CDK bootstrap ECR repository
-                  {
-                    Sid: "AllowCDKManageToolkitRepository",
-                    Effect: "Allow",
-                    Action: [
-                      "ecr:CreateRepository",
-                      "ecr:DeleteRepository",
-                      "ecr:DescribeRepositories",
-                      "ecr:PutLifecyclePolicy",
-                      "ecr:SetRepositoryPolicy",
-                    ],
-                    Resource: [
-                      "arn:aws:ecr:us-east-1:112233445566:repository/cdk-hnb659fds-container-assets-*",
-                    ],
-                  },
-                  // Permissions to create CDK bootstrap version SSM parameter -- stores version of the deployed CDK bootstrap stack, checks if it has been bootstrapped
-                  {
-                    Sid: "AllowCDKManageToolkitVersionParameter",
-                    Effect: "Allow",
-                    Action: [
-                      "ssm:DeleteParameter",
-                      "ssm:GetParameters",
-                      "ssm:PutParameter",
-                      "ssm:GetParameter",
-                    ],
-                    Resource: [
-                      "arn:aws:ssm:us-east-1:112233445566:parameter/cdk-bootstrap/hnb659fds/version",
-                    ],
-                  },
-                  // Permissions for SST to monitor the bootstrap progress.
-                  {
-                    Sid: "AllowSSTManageBootstrapStack",
-                    Effect: "Allow",
-                    Action: ["cloudformation:DescribeStacks", "cloudformation:DescribeStackEvents"],
-                    Resource: [
-                      "arn:aws:cloudformation:us-east-1:112233445566:stack/SSTBootstrap/*",
-                    ],
-                  },
-                  // Permissions for SST to assume CDK roles to deploy your application to AWS
-                  {
-                    Sid: "AllowSSTAssumeCDKToolkitRoles",
-                    Effect: "Allow",
-                    Action: "sts:AssumeRole",
-                    Resource: [
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-cfn-exec-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-file-publishing-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-image-publishing-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-lookup-role-*",
-                      "arn:aws:iam::112233445566:role/cdk-hnb659fds-deploy-role-*",
-                    ],
-                  },
-                  // Permissions for SST to monitor the deployment progress of your application
-                  {
-                    Sid: "AllowSSTMonitorStackDeployment",
-                    Effect: "Allow",
-                    Action: [
-                      "cloudformation:DeleteStack",
-                      "cloudformation:DescribeStacks",
-                      "cloudformation:DescribeStackEvents",
-                      "cloudformation:DescribeStackResources",
-                      "cloudformation:GetTemplate",
-                    ],
-                    Resource: ["arn:aws:cloudformation:us-east-1:112233445566:stack/*"],
-                  },
-                  // Permissions for SST to store metadata about your application
-                  {
-                    Sid: "AllowSSTManageBootstrapBucket",
-                    Effect: "Allow",
-                    Action: ["s3:DeleteObject", "s3:GetObject", "s3:ListBucket", "s3:PutObject"],
-                    Resource: ["arn:aws:s3:::sstbootstrap-*"],
-                  },
-                  // Permissions for SST to connect to IoT endpoint for Live Lambda development
-                  {
-                    Sid: "AllowSSTLiveLambdaSocketConnection",
-                    Effect: "Allow",
-                    Action: [
-                      "iot:DescribeEndpoint",
-                      "iot:Connect",
-                      "iot:Subscribe",
-                      "iot:Publish",
-                      "iot:Receive",
-                    ],
+                    Action: ["cloudformation:*", "s3:*", "iam:*", "lambda:*", "ecr:*", "ssm:*"],
                     Resource: ["*"],
-                  },
-                  // If you are using the RDS construct, to let SST Console to run migrations, you need the following permission
-                  {
-                    Sid: "AllowSSTCLIManageRDSMigrations",
-                    Effect: "Allow",
-                    Action: ["rds-data:ExecuteStatement"],
-                    Resource: ["arn:aws:rds:us-east-1:112233445566:cluster:*"],
-                    Condition: {
-                      Null: {
-                        "aws:ResourceTag/sst:app": "false",
-                      },
-                    },
                   },
                 ],
               },
@@ -231,12 +84,6 @@ const serverlessConfiguration: AWS = {
           ],
         },
       },
-      // SharedAuthorizer: {
-      //   Type: "AWS::ApiGateway::Authorizer",
-      //   Properties: {
-      //     Name: "auth",
-      //   },
-      // },
     },
     Outputs: {
       DeployAssumeRoleOutput: {