+++ services/environment/CHANGELOG.md
@@ -1,5 +1,11 @@
 # environment
 
+## 0.5.1
+
+### Patch Changes
+
+- @vastly/utils@0.3.2
+
 ## 0.5.0
 
 ### Minor Changes